move volunteer experience list into correct directory and format cleanup
