+++ imports/ui/views/users/volunteerExperienceList/OrganizationLogo.jsx
@@ -1,14 +1,13 @@
-import React, { Component } from 'react'
+import React, {Component} from 'react'
 
 class OrganizationLogo extends Component {
 
 	//TODO: implement clearbit to pull logo image
 	render() {
-		console.log(this.props);
 		return (
-			<div>
-				<img src={this.props.logoUrl} style={styles.img}/>
-			</div>
+				<div>
+					<img src={this.props.logoUrl} style={styles.img}/>
+				</div>
 		)
 	}
 }
@@ -16,7 +15,7 @@ class OrganizationLogo extends Component {
 const styles = {
 	img: {
 		width: '100px',
-	}
+	},
 }
 
 export default OrganizationLogo
